Casting some implicit casts making Xcode 5.1 testy
I’d change the model properties to NSIntegers, but I don’t know why there were int32_t’s in the first place

+++ Source/Models/AwfulThread.m
@@ -86,7 +86,7 @@
         info.thread = self;
         info.author = singleUser;
     }
-    info.numberOfPages = numberOfPages;
+    info.numberOfPages = (int32_t)numberOfPages;
 }
 
 #pragma mark - _AwfulThread
